geodesics now store time component aswell
