+++ src/BlackHole.jl
@@ -10,8 +10,11 @@ Includes are order sensitive
 include("geodesics.jl")
 
 include("eddingtonfinkelstein.jl")
+
 include("accretiondisk.jl")
+include("disks/geometric.jl")
+# include("disks/opticallythin.jl")
+
 include("render.jl")
 
-#export EddingtonFinkelstein, calcgeodesics, plot!, GeometricDisk, intersection, renderdisk
 end
